topics_chat_with_X/gemini_chat_with_pdf/requirements.txt
streamlit 
pillow
google-generativeai

